Improvements of the simulation
Implemented colors of cats and mice that allow us, to distinguish between cats and mice. Also the population is now divided into mice and cats with cats having their own challenge of at least eating one mouse.

+++ src/imageWriter.h
@@ -21,7 +21,7 @@ struct ImageFrameData {
     unsigned simStep;
     unsigned generation;
     std::vector<Coord> indivLocs;
-    std::vector<uint8_t> indivColors;
+    std::vector<double> indivColors;
     std::vector<Coord> barrierLocs;
     typedef std::vector<std::vector<uint8_t>> SignalLayer;  // [x][y]
     std::vector<SignalLayer> signalLayers; // [layer][x][y]
